requirements.txt
chainer
chainercv
gdown>=3.7.0
imgaug
imgviz[all]
open3d-python
opencv-python
pandas
path.py
pybullet
scikit-image
scikit-learn
scipy
termcolor
tqdm
tensorboardX
tensorflow
trimesh[easy]
